package_requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.26.0
ccxt>=2.0.0
python-binance>=1.0.16
matplotlib>=3.5.0
tabulate>=0.8.9
seaborn>=0.11.2
schedule>=1.1.0
pyinstaller>=5.6.0
python-dotenv>=0.19.0
pillow>=9.0.0
python-telegram-bot>=13.7
scipy>=1.7.0
scikit-learn>=1.0.0